+++ sam/metrics/tests/test_train_r2.py
@@ -5,7 +5,13 @@ import numpy as np
 from numpy.testing import assert_array_almost_equal
 from sam.metrics import train_r2, train_mean_r2
 
-PARAM_LIST = [np.nanmean, np.nanmedian]
+PARAM_LIST = [lambda train_data, test_n: np.nanmean(train_data),
+              lambda train_data, test_n: np.nanmedian(train_data),
+              lambda train_data, test_n: np.interp(
+                  np.arange(test_n),
+                  np.arange(len(train_data)),
+                  np.convolve(train_data, np.ones(2), mode='same')
+              )]
 
 
 class TestTrainR2(unittest.TestCase):
@@ -13,7 +19,7 @@ class TestTrainR2(unittest.TestCase):
         from sam.metrics import train_r2
         from sklearn.metrics import r2_score
 
-        for train_avg_func in PARAM_LIST:
+        for count, benchmark in enumerate(PARAM_LIST):
             with self.subTest():
                 np.random.seed(42)
                 N = 1000
@@ -35,24 +41,42 @@ class TestTrainR2(unittest.TestCase):
                     test_pred = model[train_n:]
 
                     keras_r2s.append(r2_score(test_data, test_pred))
-                    custom_r2s.append(train_r2(test_data, test_pred, train_avg_func(train_data)))
+                    custom_r2s.append(
+                        train_r2(test_data, test_pred, benchmark(train_data, test_n)))
 
                 # keras r2 should decrease with decreasing test size, custom r2 should do so less
                 assert_array_almost_equal(keras_r2s, [0.962522, 0.894608, 0.734713])
-                if train_avg_func == np.nanmean:
+                if count == 0:
                     assert_array_almost_equal(custom_r2s, [0.987602, 0.989454, 0.870157])
-                elif train_avg_func == np.nanmedian:
+                elif count == 1:
                     assert_array_almost_equal(custom_r2s, [0.987653, 0.990700, 0.938786])
+                else:
+                    assert_array_almost_equal(custom_r2s, [0.996578, 0.996330, 0.994752])
 
     def test_train_r2_shapes(self):
         # the function cannot handle data with multiple dimensions. It does however ravel
         # empty dimensions (x, 1).
-        self.assertRaises(AssertionError, train_r2, np.random.random(
-            size=(12, 2)), np.random.random(size=(12, 2)), 0)
-        self.assertRaises(AssertionError, train_r2, np.random.random(
-            size=(12, 1)), np.random.random(size=(12, 2)), 0)
-        self.assertRaises(AssertionError, train_r2, np.random.random(
-            size=(12, 2)), np.random.random(size=(12, 1)), 0)
+        with self.assertRaises(AssertionError):
+            train_r2(np.random.random(size=(12, 2)),
+                     np.random.random(size=(12, 2)),
+                     0)
+        with self.assertRaises(AssertionError):
+            train_r2(np.random.random(size=(12, 1)),
+                     np.random.random(size=(12, 2)),
+                     0)
+        with self.assertRaises(AssertionError):
+            train_r2(np.random.random(size=(12, 2)),
+                     np.random.random(size=(12, 1)),
+                     0)
+        with self.assertRaises(AssertionError):
+            train_r2(np.random.random(size=(12, 1)),
+                     np.random.random(size=(12, 1)),
+                     np.random.random(size=(12, 2)))
+        # benchmark array has to be same size as true array
+        with self.assertRaises(AssertionError):
+            train_r2(np.random.random(size=(12, 1)),
+                     np.random.random(size=(12, 1)),
+                     np.random.random(size=(13, 1)))
 
     def test_train_mean_r2_deprecation_warning(self):
         with warnings.catch_warnings(record=True) as w: